cleanup (SimpleTable): move to `~/components/Base`
