+++ components/Base/SimpleTable/SimpleTableCellSkeleton.vue
